replaced old game module
