+++ src/game/variants/mod.rs
@@ -1,1 +1,1 @@
-pub mod standard;
+pub mod basic;
