--- src/game_rewrite/mod.rs
@@ -1,6 +1,0 @@
-pub mod game;
-pub mod action;
-pub mod state;
-pub mod score;
-pub mod variants;
-pub mod error;
