--- src/game_rewrite/variants/mod.rs
@@ -1,1 +1,0 @@
-pub mod basic;
